refactor: API 레이어를 엔티티 기반 아키텍처로 변경
- src/entities/posts/ 엔티티 추가
  - index.ts: 게시물 관련 비즈니스 로직
  - types.ts: 게시물 타입 정의
- src/api/posts.ts 삭제하여 도메인 로직 분리
- 블로그 페이지에서 새로운 엔티티 import 경로 적용
