+++ src/entities/posts/types.ts
@@ -1,0 +1,10 @@
+import type { GrayMatterFile } from 'gray-matter'
+
+export interface Post extends GrayMatterFile<string> {
+  slug: string
+  data: {
+    date: string
+    title: string
+    tag: Array<string>
+  }
+}
